Ch5 revisions (#138)
**Pull Request Description**

Revises chapter 5.

Closes https://github.com/HFOSSedu/GitKit-Text/issues/94
Closes https://github.com/HFOSSedu/GitKit-Text/issues/93

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.