update 2.4.2.d hint (#173)
**Pull Request Description**

Clarify language on 2.4.2.1d hint

Closes #167

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.