DRAFT: Jan 25 sprint chap 3 (#174)
**Pull Request Description**

_Describe what your Pull Request does here_

Closes #91

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.

---------

Co-authored-by: Grant Braught <braught@dickinson.edu>
Co-authored-by: lpostner <lori.postner@ncc.edu>