Revisions to chapter 3. (#180)
**Pull Request Description**

Chapter 3 revisions by @kwurst 

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.


Co-authored-by: kwurst <karl@w-sts.com>

---------

Co-authored-by: Karl R. Wurst <karl@w-sts.com>