Some patches to chapter 4 (#182)
**Pull Request Description**

Makes some edits to chapter 4. The most significant was to revise
Section 4.3 because Codespaces automatically sets the upstream remote
when you open the cloned/forked repo.

Closes #178 

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.