remove -linux from xml:id and label values (#185)
**Pull Request Description**

A number of `xml:id` and `label` values included `-linux` which is a
holdover from the `linux` edition. The `-linux` was removed from these
values.

Closes #164 

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.