checks and updates all the condition regex (#186)
**Pull Request Description**

Checks all of the regex for:
- escaped `.` in expressions.
- inclusion of `.*` catch all.
- Some coverage of likely cases (but could still be improved).

Closes #137 
Closes #157 

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.