adds p element to introduction elements (#189)
**Pull Request Description**

Ensures that all `introduction`, `statement` and `hint` elements use
`<p>` elements.

Closes #156 

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.